requirements.txt
streamlit==1.32.0
langchain==0.1.9
openai==1.13.3
beautifulsoup4==4.12.2
requests==2.31.0
chromadb==0.4.22
python-dotenv==1.0.0
